Trienode using half zeros map size, new dictionary file created inside constructor

+++ trie/trie/extended_character_functions.hpp
@@ -6,7 +6,7 @@
 namespace trie
 {
 
-uint64_t end_of_string = 0;
+uint32_t end_of_string = 0;
 
 /* reports number of elements in the array until the first ::trie::end_of_string */
 template <typename character_t>
